requirement.txt
amqp==2.4.0
billiard==3.5.0.5
celery==4.2.1
Django==2.1.5
django-haystack==2.8.1
django-redis==4.10.0
django-tinymce==2.8.0
itsdangerous==1.1.0
jieba==0.39
kombu==4.2.2.post1
Pillow==5.4.1
py3Fdfs==2.2.0
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.0
pytz==2018.9
redis==3.1.0
vine==1.2.0
Whoosh==2.7.4
